# [DO NOT MERGE][BEAM-8151] Create 1000 SdkWorker objects instead of the current limit of 12.