--- a/sdks/python/apache_beam/runners/worker/sdk_worker_main.py
+++ b/sdks/python/apache_beam/runners/worker/sdk_worker_main.py
@@ -190,7 +190,7 @@ def _get_worker_count(pipeline_options):
   future releases.
 
   Returns:
-    an int containing the worker_threads to use. Default is 12.
+    an int containing the worker_threads to use. Default is 1000.
   """
   experiments = pipeline_options.view_as(DebugOptions).experiments
 
@@ -203,7 +203,7 @@ def _get_worker_count(pipeline_options):
           re.match(r'worker_threads=(?P<worker_threads>.*)',
                    experiment).group('worker_threads'))
 
-  return 12
+  return 1000
 
 
 def _get_state_cache_size(pipeline_options):