@lukecwik 2019-11-13
CC: @pabloem @TheNeuralBit @aaltay @robertwb @Hannah-Jiang

Follow-up for https://github.com/apache/beam/pull/9477