@lukecwik 2019-11-13
I ran the performance benchmark using the fn_api_runner_microbenchmark:

before (at commit 80e8fb59b6d785eb462e6a05b91d029fd8ff50c6):
```
run_single_pipeline, 1 element(s) at start, 100 growth per run: run 1 of 10, per element time cost: 3.0014 sec
run_single_pipeline, 1 element(s) at start, 100 growth per run: run 2 of 10, per element time cost: 0.0394958 sec
run_single_pipeline, 1 element(s) at start, 100 growth per run: run 3 of 10, per element time cost: 0.0220158 sec
run_single_pipeline, 1 element(s) at start, 100 growth per run: run 4 of 10, per element time cost: 0.0160263 sec
run_single_pipeline, 1 element(s) at start, 100 growth per run: run 5 of 10, per element time cost: 0.0130143 sec
run_single_pipeline, 1 element(s) at start, 100 growth per run: run 6 of 10, per element time cost: 0.0111442 sec
run_single_pipeline, 1 element(s) at start, 100 growth per run: run 7 of 10, per element time cost: 0.00963751 sec
run_single_pipeline, 1 element(s) at start, 100 growth per run: run 8 of 10, per element time cost: 0.00856183 sec
run_single_pipeline, 1 element(s) at start, 100 growth per run: run 9 of 10, per element time cost: 0.00796899 sec
run_single_pipeline, 1 element(s) at start, 100 growth per run: run 10 of 10, per element time cost: 0.00760294 sec


Fixed cost   3.49717886575063
Per-element  0.003790947336139101
R^2          0.963500601942131
```

after:
```
run_single_pipeline, 1 element(s) at start, 100 growth per run: run 1 of 10, per element time cost: 2.8741 sec
run_single_pipeline, 1 element(s) at start, 100 growth per run: run 2 of 10, per element time cost: 0.0401369 sec
run_single_pipeline, 1 element(s) at start, 100 growth per run: run 3 of 10, per element time cost: 0.0219585 sec
run_single_pipeline, 1 element(s) at start, 100 growth per run: run 4 of 10, per element time cost: 0.0161115 sec
run_single_pipeline, 1 element(s) at start, 100 growth per run: run 5 of 10, per element time cost: 0.0124523 sec
run_single_pipeline, 1 element(s) at start, 100 growth per run: run 6 of 10, per element time cost: 0.0108133 sec
run_single_pipeline, 1 element(s) at start, 100 growth per run: run 7 of 10, per element time cost: 0.00963711 sec
run_single_pipeline, 1 element(s) at start, 100 growth per run: run 8 of 10, per element time cost: 0.00862622 sec
run_single_pipeline, 1 element(s) at start, 100 growth per run: run 9 of 10, per element time cost: 0.00795245 sec
run_single_pipeline, 1 element(s) at start, 100 growth per run: run 10 of 10, per element time cost: 0.00736677 sec


Fixed cost   3.4638128309249883
Per-element  0.003727303158153187
R^2          0.9529920831060255
```