@lukecwik 2019-11-14
I tried generating the workers on demand as can be seen in https://github.com/lukecwik/incubator-beam/commit/c743564d3dd029260c0d8a5e703856c9ddf2650a but I ran into lots of test issues which assumed that there was only a single SdkWorker processing tasks so this is a short term solution to fix a concurrency limit issue with the Python SDK.