@lukecwik 2019-11-14
Updated perf numbers:
```
run_single_pipeline, 1 element(s) at start, 100 growth per run: run 1 of 10, per element time cost: 2.8493 sec
run_single_pipeline, 1 element(s) at start, 100 growth per run: run 2 of 10, per element time cost: 0.0399395 sec
run_single_pipeline, 1 element(s) at start, 100 growth per run: run 3 of 10, per element time cost: 0.0252568 sec
run_single_pipeline, 1 element(s) at start, 100 growth per run: run 4 of 10, per element time cost: 0.0163703 sec
run_single_pipeline, 1 element(s) at start, 100 growth per run: run 5 of 10, per element time cost: 0.0129816 sec
run_single_pipeline, 1 element(s) at start, 100 growth per run: run 6 of 10, per element time cost: 0.0109879 sec
run_single_pipeline, 1 element(s) at start, 100 growth per run: run 7 of 10, per element time cost: 0.0098083 sec
run_single_pipeline, 1 element(s) at start, 100 growth per run: run 8 of 10, per element time cost: 0.00878269 sec
run_single_pipeline, 1 element(s) at start, 100 growth per run: run 9 of 10, per element time cost: 0.00803661 sec
run_single_pipeline, 1 element(s) at start, 100 growth per run: run 10 of 10, per element time cost: 0.00757915 sec


Fixed cost   3.6176186113357542
Per-element  0.0037115805365822525
R^2          0.9044127183436924
```